add events metric

+++ application/types/stream_totals.go
@@ -3,5 +3,5 @@ package types
 type StreamTotals struct {
 	StreamName
 	ProducerId
-	EventCount int
+	EventCount float64
 }
